requirements.txt
python-telegram-bot
configparser~=5.0.2
openrouteservice~=2.3.3
